requirements.txt
absl-py==2.1.0
altair==5.5.0
astunparse==1.6.3
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
Flask==2.1.2
flatbuffers==24.3.25
fonttools==4.55.3
frozendict==2.4.6
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
html5lib==1.1
idna==3.10
imbalanced-learn==0.13.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
multitasking==0.0.11
namex==0.0.8
narwhals==1.15.1
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==10.4.0
platformdirs==4.3.6
plotly==5.3.1
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.0
scikit-learn==1.5.2
scipy==1.15.0
seaborn==0.13.2
shap==0.46.0
six==1.16.0
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.1
soupsieve==2.6
streamlit==1.39.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.0
xgboost==2.1.3
yfinance==0.2.51
